Add first backup packagers
Add current output backup methods to packagers.

Will allow to split the compression functions from the backup, and get a
generic way to store or restore a backup, all following a simple
abstract class.

Example of packagers output methods: copy the images in a directory,
pack them in a tarfile (with or without compression), compress each
image independently, etc.