Split packagers in read/write classes
As opening a write packager can sometimes overwrite the package, split
them in read/write classes: complete backups will get a readonly
packager, when pending backups will get a write packager.

Also simplify the process with tar by avoiding to temporarily open a new
readonly version of the same tarfile, as it is needed to be in readonly to
extract a file.