guard against simultaneous clean runs
When multiple clean group jobs start on several machines at the same
time and they all use the same file system for storage, file names might
exist in the global list that had vanished in the meantime and the 'open'
call will raise a FileNotFoundError exception breaking the loop.
Let's log a warning and continue processing.